+++ examples/view-integration/external-views/src/containerCode.ts
@@ -13,7 +13,7 @@ import { loadContainerRuntime } from "@fluidframework/container-runtime/legacy";
 import type { IContainerRuntime } from "@fluidframework/container-runtime-definitions/legacy";
 import type { FluidObject } from "@fluidframework/core-interfaces";
 
-import { DiceRollerFactory } from "./dataObject.js";
+import { DiceRollerFactory } from "./diceRoller.js";
 
 const diceRollerId = "dice-roller";
 const diceRollerRegistryKey = "dice-roller";
